refactor(inference): extract decoder_runtime helpers from src/cli.py
Phase A of the per-staff archival cleanup. The Stage-B decoder helpers
(_load_stage_b_crop_tensor, _LazyLogitDict, _resolve_stage_b_decode_model,
_quantize_decoder, _prepare_model_for_inference, _encode_staff_image,
_decode_stage_b_tokens) are mode-agnostic — they work for both per-staff and
per-system crops. They were trapped inside src/cli.py, which is a per-staff
module slated for archival.

Moves them to src/inference/decoder_runtime.py. Updates the only two external
import sites (src/eval/evaluate_stage_b_checkpoint.py, src/eval/tune_penalties.py)
to point at the new location. src/cli.py keeps a back-compat re-export so its
own run_stage_b path still works pending Phase B archival.

Tests: 32 torch-free tests still pass (decision_gate, musicxml_validity,
stratified_lieder_analysis, grammar_fsa_markers, vocab_marker_tokens). Code on
the runtime path (everything that imports torch) is verified by import-only
smoke checks; the GPU box runs the full integration tests.
